pyproject.toml
[tool.poetry]
name = "shopping-list"
version = "0.1.0"
description = ""
authors = ["Egor Filippov <egorfilippov@internet.ru>"]
readme = "README.md"
packages = [{include = "shopping_list"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.93.0"
sqlalchemy = "^2.0.5.post1"
pydantic = {extras = ["email"], version = "^1.10.5"}
alembic = "^1.10.1"
uvicorn = "^0.20.0"
psycopg2-binary = "^2.9.5"
asyncpg = "^0.27.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
